Refactor: Split Tab from FamPage
